Save Commit

+++ 42_cursus/CPP/cpp_09/ex01/ReversePolishNotation.hpp
@@ -7,6 +7,6 @@
 
 std::string			checkArg(std::string str);
 std::stack<char>	createStack(std::string str);
-int					printResult(std::stack<char> * myStack);
+double				printResult(std::stack<char> * myStack);
 
 #endif
